fixed a glaring bug in the pyramid algorithm that a cs 143 ta would crucify you for.
Specifically, it was a quadruply nested loop (actually eight times nested, but we are only interested in the four right now) where I did not reset the iterators for the inner loops during each iteration of the outer loop. This not only caused the pyramid to miss lots of valid pyramids, but also caused it to think that some bad pyramids were unique because it would not detect the other pyramid, which is far worse!

Also found and fixed bugs with spherical coordinate conversion.