Add generation centroid mag cutoff, split apart input and expected stars.
Needed to split them up because the star-ids from the algos should be compared against expected stars and everything.

Still a bit iffy: Even though we have oracle knowledge about which input stars correspond to which expected stars, we do a brute force nearest check in StarIdsCompare. Theoretically I think this is fine? After all, if there are two centroids so close to each other that perturbations cause them to become closer to other expected stars, then it's only reasonable to expect that the star-id will be different too...