Rewrite most of how serialization/deserialization works.
For the following reasons:

(a), to ensure that values are aligned, which is required for direct reading on many CPU
architectures.

(b), to avoid the duality of Serialize* and Serialize*Length functions.

(c), to overall reduce the room for error, by providing [de]serialization template functions that
abstract away the pointer wrangling.